Fix error on mega bootup
After loading this script into the Mega and issuing a coldboot, the
code fails to work if this is set to "SPI_FULL_SPEED" and instead hangs
after sending the "index.htm" file, there is also additional corruption
of sent files.
the code functions correctly if the "reset" button is pressed.
Appears to be an issue in the SPI library