Implemented a _DEBUG flag for serial debugging
This library is an Ethernet webserver, it should not 'corrupt' the
serial port with debugging messages by default.
There is a risk that a noob who implements a serial port in  his own
sketch code would get incorrect output. They may be unable to correctly
comment out the debug code in the library